+++ g/modules/constants.js
@@ -1,0 +1,3 @@
+export const MAX_DIGITS = 10;
+export const MAX_NUMBER = 1e20;
+export const MIN_NUMBER = -1e20;
